--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,4 @@
 module github.com/buger/jsonparser
 
 go 1.13
+